refactor: package naming을 ourboard로 수정
