--- src/main/java/com/twentiethcenturygangsta/jamboard/repository/AuthenticatedAdminMemberRepository.java
@@ -1,9 +1,0 @@
-package com.twentiethcenturygangsta.jamboard.repository;
-
-import com.twentiethcenturygangsta.jamboard.entity.AuthenticatedAdminMember;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AuthenticatedAdminMemberRepository extends JpaRepository<AuthenticatedAdminMember, Long> {
-}
