+++ src/main/java/com/twentiethcenturygangsta/ourboard/JamBoardApplication.java
@@ -1,4 +1,4 @@
-package com.twentiethcenturygangsta.jamboard;
+package com.twentiethcenturygangsta.ourboard;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
