+++ src/main/java/com/twentiethcenturygangsta/ourboard/auth/Role.java
@@ -1,4 +1,4 @@
-package com.twentiethcenturygangsta.jamboard.auth;
+package com.twentiethcenturygangsta.ourboard.auth;
 
 public enum Role {
     SUPER_USER, DEFAULT_USER
