+++ src/main/java/com/twentiethcenturygangsta/ourboard/config/MvcConfiguration.java
@@ -1,4 +1,4 @@
-package com.twentiethcenturygangsta.jamboard.config;
+package com.twentiethcenturygangsta.ourboard.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
