+++ src/main/java/com/twentiethcenturygangsta/ourboard/dto/TablesInfo.java
@@ -1,9 +1,7 @@
-package com.twentiethcenturygangsta.jamboard.dto;
+package com.twentiethcenturygangsta.ourboard.dto;
 
 import lombok.Builder;
 import lombok.Getter;
-
-import java.util.ArrayList;
 
 @Getter
 public class TablesInfo {
