+++ src/main/java/com/twentiethcenturygangsta/ourboard/site/AuthMember.java
@@ -1,4 +1,4 @@
-package com.twentiethcenturygangsta.jamboard.site;
+package com.twentiethcenturygangsta.ourboard.site;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
