+++ src/main/java/com/twentiethcenturygangsta/ourboard/trace/JamBoardEntityAspect.java
@@ -1,4 +1,4 @@
-package com.twentiethcenturygangsta.jamboard.trace;
+package com.twentiethcenturygangsta.ourboard.trace;
 
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.annotation.Aspect;
@@ -10,7 +10,7 @@ import java.lang.reflect.Method;
 @Aspect
 public class JamBoardEntityAspect {
 
-    @Before("@annotation(com.twentiethcenturygangsta.jamboard.trace.JamBoardEntity)")
+    @Before("@annotation(com.twentiethcenturygangsta.ourboard.trace.JamBoardEntity)")
     public String getGroup(JoinPoint joinPoint) {
         MethodSignature methodSignature = (MethodSignature) joinPoint.getSignature();
         Method method = methodSignature.getMethod();
