+++ src/test/java/com/twentiethcenturygangsta/ourboard/JamBoardApplicationTests.java
@@ -1,4 +1,4 @@
-package com.twentiethcenturygangsta.jamboard;
+package com.twentiethcenturygangsta.ourboard;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
